Breaking: Use presigned URLs from S3 for downloads
This to avoid having to proxy all downloads through filebin, which was too much for filebin.net to handle. A relatively unfortunate consequence is that server side encryption is disabled since decryption of presigned downloads is not possible. This is a breaking change since the nonce column as been removed from the file table in the database schema.

In order to upgrade existing installations, it is possible to drop the nonce column from the file table manually using: ALTER TABLE file DROP COLUMN nonce;